Put authentication string method into utils, as several modules need it.
