+++ b2handle/handleclient.py
@@ -10,11 +10,9 @@ from handleexceptions import *
 import hsresponses
 from handlesystemconnector import HandleSystemConnector
 import requests
-import urllib
 import json
 import copy
 import xml.etree.ElementTree as ET
-import base64
 import uuid
 import logging
 import re
@@ -1192,27 +1190,6 @@ class EUDATHandleClient(object):
                 indices.append(entry['index'])
         return indices
 
-    @staticmethod
-    def create_authentication_string(username, password):
-        '''
-        Create an authentication string from the username and password.
-
-        :param username: Username.
-        :param password: Password.
-        :return: The encoded string.
-        '''
-
-        LOGGER.debug('create_authentication_string...')
-
-        username_utf8 = username.encode('utf-8')
-        userpw_utf8 = password.encode('utf-8')
-        username_perc = urllib.quote(username_utf8)
-        userpw_perc = urllib.quote(userpw_utf8)
-
-        authinfostring = username_perc + ':' + userpw_perc
-        authinfostring_base64 = base64.b64encode(authinfostring)
-        return authinfostring_base64
-
     def check_if_username_exists(self, username):
         '''
         Check if the username handles exists.
@@ -1406,7 +1383,7 @@ class EUDATHandleClient(object):
         :param password: The password contained in the index of the username
             handle.
         '''
-        auth = self.create_authentication_string(username, password)
+        auth = util.create_authentication_string(username, password)
         self.__HS_auth_string = auth
 
     def __set_revlookup_auth_string(self, username, password):
@@ -1418,7 +1395,7 @@ class EUDATHandleClient(object):
         :param username: Username.
         :param password: Password.
         '''
-        auth = self.create_authentication_string(username, password)
+        auth = util.create_authentication_string(username, password)
         self.__revlookup_auth_string = auth
 
     def __get_handle_record_if_necessary(self, handle, handlerecord_json):
